3.1.1. przeniesienie implementacji stosu z klasy DefaultCountingOutRhymer do klasy IntArrayStack
